// src/LiFi/types.ts
export interface ILifiGasSuggestionParams {
  fromToken?: string;
  fromChain?: string | number;
  toChain?: string | number;
  toToken?: string;
}

/** Represents a token within the Li.FI API context. */
export interface ILifiToken {
  address: string;
  decimals: number;
  symbol: string;
  chainId: number;
  coinKey?: string;
  name: string;
  logoURI?: string;
  priceUSD?: string;
}

/** Represents bid data, potentially related to RFQ or specific protocols used internally by Li.FI. */
export interface ILifiBid {
  user: string;
  router: string;
  initiator: string;
  sendingChainId: number;
  sendingAssetId: string;
  amount: string;
  receivingChainId: number;
  receivingAssetId: string;
  amountReceived: string;
  receivingAddress: string;
  transactionId: string;
  expiry: number;
  callDataHash: string;
  callTo: string;
  encryptedCallData: string;
  sendingChainTxManagerAddress: string;
  receivingChainTxManagerAddress: string;
  bidExpiry: number;
  bidSignature: string;
  gasFeeInReceivingToken: string;
  totalFee: string;
  metaTxRelayerFee: string;
  routerFee: string;
}

/** Wrapper for bid data within a step. */
export interface ILifiData {
  bid: ILifiBid;
}

/** Represents estimated costs and outcomes for a Li.FI swap step or route. */
export interface ILifiEstimate {
  fromAmount: string;
  toAmount: string;
  toAmountMin: string;
  approvalAddress: string;
  feeCosts: IFeeCost[];
  gasCosts: IGasCost[];
}

/** Represents the core action within a Li.FI step (e.g., swap, bridge). */
export interface ILifiAction {
  fromChainId: number;
  fromAmount: string;
  toChainId: number;
  fromToken: ILifiToken;
  toToken: ILifiToken;
  slippage?: number;
  fromAddress: string;
  toAddress: string;
}

/** Represents a single step within a Li.FI route. */
export interface ILifiSwapStep {
  id: string;
  type: string;
  toolDetails: IToolDetails;
  tool: string;
  action: ILifiAction;
  data?: ILifiData;
  integrator?: string;
  referrer?: string;
  execution?: string;
  estimate: ILifiEstimate;
  transactionRequest?: TransactionRequest;
}
// ...
export interface ILifiBestQuote extends ILifiSwapStep {
  transactionRequest: TransactionRequest;
  estimate: ILifiEstimate;
  includedSteps?: ILifiSwapStep[];
}

/** Detailed token information for transaction status. */
export interface ILifiTokenDetails {
  address: string;
  chainId: number;
  symbol: string;
  decimals: number;
  name: string;
  coinKey: string;
  logoURI: string;
  priceUSD: string;
}

/** Details about the sending side of a Li.FI transaction. */
export interface ILifiSendingDetails {
  txHash: string;
  txLink: string;
  amount: string;
  token: ILifiTokenDetails;
  chainId: number;
  gasPrice: string;
  gasUsed: string;
  gasToken: ILifiTokenDetails;
  gasAmount: string;
  gasAmountUSD: string;
  amountUSD: string;
  value: string;
  timestamp: number;
}

/** Details about the receiving side of a Li.FI transaction. */
export interface ILifiReceivingDetails {
  txHash: string;
  txLink: string;
  amount: string;
  token: ILifiTokenDetails;
  chainId: number;
  gasPrice: string;
  gasUsed: string;
  gasToken: ILifiTokenDetails;
  gasAmount: string;
  gasAmountUSD: string;
  amountUSD: string;
  value: string;
  timestamp: number;
}

/** Metadata associated with a Li.FI transaction. */
export interface ILifiMetadata {
  integrator: string;
}

/** Represents the status of a Li.FI transaction. */
export interface ILifiTransactionStatus {
  transactionId: string;
  sending: ILifiSendingDetails;
  receiving: ILifiReceivingDetails;
  lifiExplorerLink: string;
  fromAddress: string;
  toAddress: string;
  tool: string;
  status: string;
  substatus: string;
  substatusMessage: string;
  metadata: ILifiMetadata;
  bridgeExplorerLink: string;
}